feat(cache): implement cache with redis for several service

+++ src/services/postgres/PlaylistService.js
@@ -3,6 +3,7 @@ const { Pool } = require('pg');
 const InvariantError = require('../../exceptions/InvariantError');
 const NotFoundError = require('../../exceptions/NotFoundError');
 const AuthorizationError = require('../../exceptions/AuthorizationError');
+const cacheKeys = require('../../utils/cacheKeys');
 
 class PlaylistService {
   /**
@@ -11,9 +12,10 @@ class PlaylistService {
    * 2. Get
    * 3. Delete (id)
    */
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool();
     this._collaborationService = collaborationService;
+    this._cacheService = cacheService;
   }
 
   async addPlaylist(name, owner) {
@@ -30,29 +32,48 @@ class PlaylistService {
       throw new InvariantError('Playlist has been failed to create');
     }
 
+    //Cache Redis
+    await this._cacheService.delete(cacheKeys.playlists.playlistByUser(owner));
+
     return result.rows[0].id;
   }
 
   async getPlaylists(owner) {
-    // Use Join table to get username
-    const query = {
-      text: `SELECT playlists.id, playlists.name, users.username
-      FROM playlists
-      LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
-      LEFT JOIN users ON users.id = playlists.owner
-      WHERE playlists.owner = $1 OR collaborations.user_id = $1
-      GROUP BY playlists.id, playlists.name, users.username`,
-      values: [owner],
-    };
+    try {
+      //Cache Redis
+      const result = await this._cacheService.get(
+        cacheKeys.playlists.playlistByUser(owner)
+      );
 
-    const result = await this._pool.query(query);
+      return JSON.parse(result);
+      // eslint-disable-next-line no-unused-vars
+    } catch (error) {
+      // Use Join table to get username
+      const query = {
+        text: `SELECT playlists.id, playlists.name, users.username
+        FROM playlists
+        LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
+        LEFT JOIN users ON users.id = playlists.owner
+        WHERE playlists.owner = $1 OR collaborations.user_id = $1
+        GROUP BY playlists.id, playlists.name, users.username`,
+        values: [owner],
+      };
 
-    return result.rows;
+      const result = await this._pool.query(query);
+
+      //Cache Redis
+      await this._cacheService.set(
+        cacheKeys.playlists.playlistByUser(owner),
+        JSON.stringify(result.rows),
+        600
+      );
+      return result.rows;
+    }
   }
 
   async deletePlaylistById(id) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
@@ -63,11 +84,15 @@ class PlaylistService {
         'Failed to delete playlists. PlaylistsId not found'
       );
     }
+    const { owner } = result.rows[0];
+
+    //Cache Redis
+    await this._cacheService.delete(cacheKeys.playlists.playlistByUser(owner));
   }
 
   async verifyPlaylistOwner(id, owner) {
     const query = {
-      text: 'SELECT * FROM playlists WHERE id = $1',
+      text: 'SELECT owner FROM playlists WHERE id = $1',
       values: [id],
     };
 
@@ -123,24 +148,47 @@ class PlaylistService {
 
     const action = 'add';
     await this.addPlaylistActivity(playlistId, songId, userId, action);
+
+    //Cache Redis
+    await this._cacheService.delete(
+      cacheKeys.playlists.playlistSongById(playlistId)
+    );
   }
 
   //GET ALL
   async getAllPlaylistSong(playlistId) {
-    const playlistData = await this.getPlaylistById(playlistId);
+    try {
+      const result = JSON.parse(
+        await this._cacheService.get(
+          cacheKeys.playlists.playlistSongById(playlistId)
+        )
+      );
 
-    const query = {
-      text: `SELECT songs.id, songs.title, songs.performer
-		          FROM playlist_songs
-		          RIGHT JOIN songs ON songs.id = playlist_songs.song_id
-		          WHERE playlist_songs.playlist_id = $1`,
-      values: [playlistId],
-    };
+      return result;
+      // eslint-disable-next-line no-unused-vars
+    } catch (error) {
+      const playlistData = await this.getPlaylistById(playlistId);
 
-    const songsData = await this._pool.query(query);
-    playlistData.songs = songsData.rows;
+      const query = {
+        text: `SELECT songs.id, songs.title, songs.performer
+                FROM playlist_songs
+                RIGHT JOIN songs ON songs.id = playlist_songs.song_id
+                WHERE playlist_songs.playlist_id = $1`,
+        values: [playlistId],
+      };
 
-    return playlistData;
+      const songsData = await this._pool.query(query);
+      playlistData.songs = songsData.rows;
+
+      //Cache Redis
+      await this._cacheService.set(
+        cacheKeys.playlists.playlistSongById(playlistId),
+        JSON.stringify(playlistData),
+        600
+      );
+
+      return playlistData;
+    }
   }
 
   //DELETE
@@ -158,6 +206,11 @@ class PlaylistService {
 
     const action = 'delete';
     await this.addPlaylistActivity(playlistId, songId, userId, action);
+
+    //Cache Redis
+    await this._cacheService.delete(
+      cacheKeys.playlists.playlistSongById(playlistId)
+    );
   }
 
   /**
@@ -173,23 +226,44 @@ class PlaylistService {
     };
 
     await this._pool.query(query);
+
+    //Cache Redis
+    await this._cacheService.delete(
+      cacheKeys.playlists.playlistActivityById(playlistId)
+    );
   }
 
   async getPlaylistActivityByPlaylistId(playlistId) {
-    const query = {
-      text: `SELECT users.username, songs.title,
-                    psa.action, psa.time
-	           FROM playlist_song_activities psa
-             LEFT JOIN users ON users.id = psa.user_id
-             LEFT JOIN songs ON songs.id = psa.song_id
-             WHERE psa.playlist_id = $1
-             ORDER BY time`,
-      values: [playlistId],
-    };
+    try {
+      const result = await this._cacheService.get(
+        cacheKeys.playlists.playlistActivityById(playlistId)
+      );
 
-    const result = await this._pool.query(query);
+      return JSON.parse(result);
+      // eslint-disable-next-line no-unused-vars
+    } catch (error) {
+      const query = {
+        text: `SELECT users.username, songs.title,
+                      psa.action, psa.time
+               FROM playlist_song_activities psa
+               LEFT JOIN users ON users.id = psa.user_id
+               LEFT JOIN songs ON songs.id = psa.song_id
+               WHERE psa.playlist_id = $1
+               ORDER BY time`,
+        values: [playlistId],
+      };
 
-    return result.rows;
+      const result = await this._pool.query(query);
+
+      //Cache Redis
+      await this._cacheService.set(
+        cacheKeys.playlists.playlistActivityById(playlistId),
+        JSON.stringify(result.rows),
+        300
+      );
+
+      return result.rows;
+    }
   }
 
   //Collaborations feature service need
